refactor(app): canonicalize totp app entrypoint module paths
