+++ packages/app/2fa-setup.html
@@ -20,6 +20,6 @@
   <body class="antialiased overscroll-none text-12-regular overflow-hidden">
     <noscript>You need to enable JavaScript to run this app.</noscript>
     <div id="root"></div>
-    <script src="/src/2fa-setup/index.tsx" type="module"></script>
+    <script src="/src/totp-setup/index.tsx" type="module"></script>
   </body>
 </html>
