+++ packages/app/src/2fa-setup/setup.tsx
@@ -1,1 +1,1 @@
-export { TotpSetupApp, TwoFactorSetupApp } from "@opencode-ai/fork-ui"
+export { TotpSetupApp, TwoFactorSetupApp } from "../totp-setup/setup"
