+++ packages/app/src/2fa/index.tsx
@@ -1,13 +1,2 @@
-// @refresh reload
-import { render } from "solid-js/web"
-import { TotpApp } from "./verify"
-import "@/index.css"
-
-const root = document.getElementById("root")
-if (import.meta.env.DEV && !(root instanceof HTMLElement)) {
-  throw new Error(
-    "Root element not found. Did you forget to add it to your 2fa.html? Or maybe the id attribute got misspelled?",
-  )
-}
-
-render(() => <TotpApp />, root!)
+// Legacy entrypoint shim. Prefer /src/totp/index.tsx.
+import "../totp/index"
