+++ packages/app/src/2fa/verify.tsx
@@ -1,1 +1,1 @@
-export { TotpApp, TwoFactorApp } from "@opencode-ai/fork-ui"
+export { TotpApp, TwoFactorApp } from "../totp/verify"
