+++ packages/app/src/totp-setup/setup.tsx
@@ -1,0 +1,1 @@
+export { TotpSetupApp, TwoFactorSetupApp } from "@opencode-ai/fork-ui"
