+++ packages/app/src/totp/index.tsx
@@ -1,0 +1,13 @@
+// @refresh reload
+import { render } from "solid-js/web"
+import { TotpApp } from "./verify"
+import "@/index.css"
+
+const root = document.getElementById("root")
+if (import.meta.env.DEV && !(root instanceof HTMLElement)) {
+  throw new Error(
+    "Root element not found. Did you forget to add it to your totp.html (or legacy 2fa.html)? Or maybe the id attribute got misspelled?",
+  )
+}
+
+render(() => <TotpApp />, root!)
